requirements.txt
speechrecognition
gtts
playsound
python-docx
